+++ email/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added `OAuth2Config::redirect_host` and `OAuth2Config::redirect_port` so that they can be customized.
 - Added new boolean `Envelope::has_attachment` to determine if an envelope has at least one attachment.
 - Added `ImapConfig::extensions` of type `Option<ImapExtensionsConfig>`.
 - Changed the way `ID` command is automatically sent after authentication.