+++ email/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.0] - 2024-10-28
+
 ### Added
 
 - Added `tokio-rustls` cargo feature. The aim is to be able, one day, to compile `email-lib` with other async runtimes and other TLS options.